Remove setup.py indirection for DB related actions
We do a number of things to avoid just calling a script normally:

 * Have a setup.py
 * Install the package in the dev env
 * This now forces us to compile the dependencies as well

This results in about a 6 second delay when starting fresh, and mostly just makes how this works more confusing.
