+++ tox.ini
@@ -44,7 +44,6 @@ passenv =
     dev: GOOGLE_CLIENT_SECRET
     tests: TEST_DATABASE_URL
 deps =
-    dev: -e .
     dev: -r requirements/dev.txt
     tests: -r requirements/tests.txt
     functests: -r requirements/functests.txt
